Still working with multiple clients, but now GLFW and ENET are linked statically. Soon instructions and maybe even building scripts.

+++ core/networking/client.cpp
@@ -15,8 +15,8 @@ namespace Networking
 	ENetHost *cclient = NULL;
 	ENetPeer *cserver = NULL;
 	
-	int send_string_to_server(const char*string,int id);
-	int send_vec2_to_server(float x, float y,int id) {
+	int send_string_to_server(const char*string,uint8_t id);
+	int send_vec2_to_server(float x, float y,uint8_t id) {
 		if (cserver == NULL)
 			return -1;
 		packet_t pkt;
@@ -24,7 +24,7 @@ namespace Networking
 		pkt.payload.data.yf=y;
 		pkt.type=packet_types::vec2_packet;
 		pkt.size=sizeof(pkt);
-		pkt.id=static_cast<uint8_t>(id);
+		pkt.id=id;
 		send_packet(cserver, &pkt,pkt.size,false);
 //		enet_host_flush(cclient); NOTE: should I flush it every time?
 		return 0;
